+++ docs/source/release-notes.rst
@@ -15,7 +15,7 @@ v1.9.4 (2025-xx-xx)
   list.
 - Fix a bug in line model that causes incorrect admittance matrix.
 - Correct a typo in the PSSE DYR parser, changing "Tprod" to "Tpord".
-- Add a TurbineGov model ``IEEEG1NL``
+- Add TurbineGov models ``IEEEG1GV`` and ``IEEEG1TS``
 
 v1.9.3 (2025-01-05)
 -------------------